splitcloud/AWS/utils.py
from storages.backends.s3boto3 import S3Boto3Storage

StaticRootS3BotoStorage = lambda: S3Boto3Storage(location='static')
MediaRootS3BotoStorage  = lambda: S3Boto3Storage(location='media')

class ProtectedS3Storage(S3Boto3Storage):
    location='protected'
