requirements.txt
boto==2.49.0
boto3==1.9.189
botocore==1.12.189
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-storages==1.7.1
docutils==0.14
idna==2.8
jmespath==0.9.4
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
stripe==2.32.1
urllib3==1.25.3
wincertstore==0.2
